feat(hq-inbox): 整合「異常處理」進收件匣 ⚠️ 異常 tab
把 /wms/exceptions 整頁收進總倉收件匣的 source chip,跟 撿貨單 /
補貨申請 / 轉貨單 / 互助訂單 / 短少訂單 並列。

改動:
- 新增 components/ExceptionsContent.tsx — 抽離原 /wms/exceptions 頁的
  資料抓取 + 表格 + tabs + TransferShortageResolveModal 邏輯為獨立元件
  - props: showHeader (預設 true)、onCountChange callback 回報總筆數
- hq/inbox/page.tsx:
  - SourceTag union 加 "exception",所有 Record<SourceTag,...> 同步補
  - chip bar 加「⚠️ 異常」chip(rose 色)
  - sourceFilter === "exception" → render <ExceptionsContent showHeader=false />
    取代既有的 stage tab / batch / row list / pagination
  - exceptionCount 透過 onCountChange 從元件回報,chip badge 用它
    (跟其他 source 不同 — 異常不走 rpc_inbox_counts)
  - URL ?source=exception / localStorage 都支援
- layout.tsx:
  - 拿掉 sidebar「異常處理」item + GUARDED_PATHS /wms/exceptions
- wms/exceptions/page.tsx → redirect 到 /hq/inbox?source=exception
  (舊書籤兼容)
