+++ apps/admin/src/app/(protected)/layout.tsx
@@ -51,7 +51,6 @@ const NAV: NavGroup[] = [
       { href: "/wms/receiving", label: "進貨待辦", match: /^\/wms\/receiving/ },
       { href: "/wms/picking", label: "派貨工作台", match: /^\/wms\/picking(?!\/history)|^\/picking\/workstation/ },
       { href: "/wms/transfers", label: "內部調撥", match: /^\/wms\/transfers|^\/transfers\/dispatch|^\/transfers$|^\/transfers\/?$/ },
-      { href: "/wms/exceptions", label: "異常處理", match: /^\/wms\/exceptions/ },
     ],
   },
   {
@@ -81,7 +80,6 @@ const BRANCH_HIDDEN_HREFS = new Set([
   "/wms/receiving",       // 進貨待辦 (HQ)
   "/wms/picking",         // 派貨工作台
   "/wms/transfers",       // 內部調撥 (HQ-side)
-  "/wms/exceptions",      // 異常處理 (HQ)
   // /wms/inbound 不在這裡 — 分店要用,屬於分店業務
   "/finance/receivables", // HQ 應收
 ]);
